nc_image_reader -> reading; better file structure; easier subclassing
