+++ setup.cfg
@@ -79,8 +79,8 @@ testing =
 
 [options.entry_points]
 console_scripts =
-    repurpose_images = qa4sm_preprocessing.nc_image_reader.cli:run_repurpose
-    transpose_images = qa4sm_preprocessing.nc_image_reader.cli:run_transpose
+    repurpose_images = qa4sm_preprocessing.reading.cli:run_repurpose
+    transpose_images = qa4sm_preprocessing.reading.cli:run_transpose
 
 [tool:pytest]
 # Specify command line options as you would do when invoking pytest directly.
