--- src/qa4sm_preprocessing/nc_image_reader/__init__.py
@@ -1,11 +1,0 @@
-# -*- coding: utf-8 -*-
-from pkg_resources import get_distribution, DistributionNotFound
-
-try:
-    # Change here if project is renamed and does not equal the package name
-    dist_name = __name__
-    __version__ = get_distribution(dist_name).version
-except DistributionNotFound:  # pragma: no cover
-    __version__ = 'unknown'
-finally:
-    del get_distribution, DistributionNotFound
