+++ src/qa4sm_preprocessing/reading/utils.py
@@ -1,6 +1,6 @@
 import datetime
 from typing import Sequence
-from math import floor, sqrt
+from math import floor
 import numpy as np
