+++ tests/test_reading/test_cli.py
@@ -5,8 +5,8 @@ import shutil
 import netCDF4
 import xarray as xr
 
-from qa4sm_preprocessing.nc_image_reader.readers import GriddedNcOrthoMultiTs, XarrayTSReader
-from qa4sm_preprocessing.nc_image_reader.cli import repurpose, transpose
+from qa4sm_preprocessing.reading import GriddedNcOrthoMultiTs, XarrayTSReader
+from qa4sm_preprocessing.reading.cli import repurpose, transpose
 
 # this is defined in conftest.py
 from pytest import test_data_path
@@ -28,13 +28,11 @@ def cli_args_lis(test_output_path):
         *("--latdim", "north_south"),
         *("--londim", "east_west"),
         *("--level", "SoilMoist_profiles:0"),
-        *("--lat", "29.875", "0.25"),
-        *("--lon", "-11.375", "0.25"),
+        *("--lat", "29.875", "54.75", "0.25"),
+        *("--lon", "-11.375", "1.0", "0.25"),
     ]
 
-# @pytest.mark.skip(
-#     reason="Function nc_image_reader.transpose._traspose is not compatible and should be updated"
-# )
+
 def test_transpose_lis(cli_args_lis, lis_noahmp_stacked):
     args = cli_args_lis
     args[1] = args[1] + "/lis_noahmp_transposed.nc"
